Adding toString() to print City objects
Adding the toString() will simplify displaying City objects. Take for example
when a ListView with an ArrayAdapter<City> is used to display city results.
The City's toString() method will be called to convert each item in the array
into a view. The toString() will streamline the process of displaying City results
in a view.